Make room for adding more colorspaces to ffdshow
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@3943 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/TffdshowEnc.h
@@ -125,7 +125,7 @@ protected:
     unsigned int dx,dy;
     unsigned int outDx,outDy;
     Tcsps enccsps;
-    int enccsp;
+    uint64_t enccsp;
     Tconvert *convert;
     int keyspacing;
     int getQuantQuant(void);
@@ -137,7 +137,7 @@ protected:
     bool firstrun;
 
     HWND h_graph;
-    int inColorspace;
+    uint64_t inColorspace;
     BITMAPINFO biOutput;
     TencStats encStats;
