+++ src/codecs/TvideoCodecLibavcodec.h
@@ -77,7 +77,7 @@ protected:
     bool mpeg2_new_sequence;
     bool bReorderBFrame;
 protected:
-    virtual LRESULT beginCompress(int cfgcomode,int csp,const Trect &r);
+    virtual LRESULT beginCompress(int cfgcomode,uint64_t csp,const Trect &r);
     virtual bool beginDecompress(TffPictBase &pict,FOURCC infcc,const CMediaType &mt,int sourceFlags);
     virtual HRESULT flushDec(void);
 public:
