+++ src/convert/Tconvert.h
@@ -11,7 +11,7 @@
 //#define AVISYNTH_BITBLT //use avisynth bitblt function to just copy frame when no colorspace conversion is needed
 #define XVID_BITBLT //use xvid's YV12 -> YV12 copy function - seems to be fastest
 
-static __inline bool outcsp_sup_ffdshow_converter(int outcsp)
+static __inline bool outcsp_sup_ffdshow_converter(uint64_t outcsp)
 {
     return (Tconfig::cpu_flags&FF_CPU_SSE2) && ((outcsp & FF_CSPS_MASK) & (FF_CSP_RGB24|FF_CSP_RGB32|FF_CSP_BGR24|FF_CSP_BGR32));
 }
@@ -33,8 +33,8 @@ private:
     Tlibavcodec *libavcodec;
     Tswscale *swscale;
     bool initsws;
-    int oldincsp,oldoutcsp;
-    int incsp1,outcsp1;
+    uint64_t oldincsp,oldoutcsp;
+    uint64_t incsp1,outcsp1;
     const TcspInfo *incspInfo,*outcspInfo;
     int rgbInterlaceMode;
     TffdshowConverters *ffdshow_converters;
@@ -80,17 +80,17 @@ public:
     Tconvert(Tlibavcodec *Ilibavcodec,bool highQualityRGB,unsigned int Idx,unsigned int Idy,const TrgbPrimaries &IrgbPrimaries, int rgbInterlaceMode, bool dithering);
     ~Tconvert();
     unsigned int dx,dy,outdy;
-    int convert(int incsp,
+    int convert(uint64_t incsp,
                 const uint8_t*const src[],
                 const stride_t srcStride[],
-                int outcsp,
+                uint64_t outcsp,
                 uint8_t* dst[],
                 stride_t dstStride[],
                 const Tpalette *srcpal,
                 int video_full_range_flag = VIDEO_FULL_RANGE_INVALID,
                 int YCbCr_RGB_matrix_coefficients = YCbCr_RGB_coeff_Unspecified,
                 bool vram_indirect=false);
-    int convert(const TffPict &pict,int outcsp,uint8_t* dst[],stride_t dstStride[],bool vram_indirect=false);
+    int convert(const TffPict &pict,uint64_t outcsp,uint8_t* dst[],stride_t dstStride[],bool vram_indirect=false);
 };
 
 class TffColorspaceConvert :public CUnknown,
@@ -108,10 +108,10 @@ public:
     TffColorspaceConvert(LPUNKNOWN punk,HRESULT *phr);
     virtual ~TffColorspaceConvert();
 
-    STDMETHODIMP allocPicture(int csp,unsigned int dx,unsigned int dy,uint8_t *data[],stride_t stride[]);
+    STDMETHODIMP allocPicture(uint64_t csp,unsigned int dx,unsigned int dy,uint8_t *data[],stride_t stride[]);
     STDMETHODIMP freePicture(uint8_t *data[]);
-    STDMETHODIMP convert(unsigned int dx,unsigned int dy,int incsp,uint8_t *src[],const stride_t srcStride[],int outcsp,uint8_t *dst[],stride_t dstStride[]);
-    STDMETHODIMP convertPalette(unsigned int dx,unsigned int dy,int incsp,uint8_t *src[],const stride_t srcStride[],int outcsp,uint8_t *dst[],stride_t dstStride[],const unsigned char *pal,unsigned int numcolors);
+    STDMETHODIMP convert(unsigned int dx,unsigned int dy,uint64_t incsp,uint8_t *src[],const stride_t srcStride[],uint64_t outcsp,uint8_t *dst[],stride_t dstStride[]);
+    STDMETHODIMP convertPalette(unsigned int dx,unsigned int dy,uint64_t incsp,uint8_t *src[],const stride_t srcStride[],uint64_t outcsp,uint8_t *dst[],stride_t dstStride[],const unsigned char *pal,unsigned int numcolors);
 };
 
 #endif
