+++ src/convert/ffdshow_converters.h
@@ -61,8 +61,8 @@ typedef ptrdiff_t stride_t;
 class TffdshowConverters
 {
 public:
-    void init(int incsp,                      // FF_CSP_420P, FF_CSP_NV12, FF_CSP_YUY2 or FF_CSP_420P (progressive only)
-              int outcsp,                     // FF_CSP_RGB32, FF_CSP_RGB24, FF_CSP_BGR32 or FF_CSP_BGR24
+    void init(uint64_t incsp,                 // FF_CSP_420P, FF_CSP_NV12, FF_CSP_YUY2 or FF_CSP_420P (progressive only)
+              uint64_t outcsp,                // FF_CSP_RGB32, FF_CSP_RGB24, FF_CSP_BGR32 or FF_CSP_BGR24
               ffYCbCr_RGB_MatrixCoefficientsType cspOptionsIturBt,  // ffYCbCr_RGB_coeff_ITUR_BT601, ffYCbCr_RGB_coeff_ITUR_BT709 or ffYCbCr_RGB_coeff_SMPTE240M
               int input_Y_white_level,        // input Y level (TV:235, PC:255)
               int input_Y_black_level,        // input Y level (TV:16, PC:0)
@@ -88,7 +88,7 @@ public:
     ~TffdshowConverters();
 
 private:
-    int m_incsp, m_outcsp;
+    uint64_t m_incsp, m_outcsp;
     int m_thread_count;
     bool m_rgb_limit;
     bool m_dithering;
@@ -109,7 +109,7 @@ private:
         __m128i cB_Cb;
     } *m_coeffs;
 
-    template<int incsp, int outcsp, int left_edge, int right_edge, int rgb_limit, int aligned, bool dithering> static __forceinline
+    template<uint64_t incsp, uint64_t outcsp, int left_edge, int right_edge, int rgb_limit, int aligned, bool dithering> static __forceinline
     void convert_a_unit(const unsigned char* &srcY,
                         const unsigned char* &srcCb,
                         const unsigned char* &srcCr,
@@ -132,7 +132,7 @@ private:
         stride_t stride_CbCr,
         stride_t stride_dst);
 
-    template <int incsp, int rgb_limit> void convert_translate_outcsp(
+    template <uint64_t incsp, int rgb_limit> void convert_translate_outcsp(
         const uint8_t* srcY,
         const uint8_t* srcCb,
         const uint8_t* srcCr,
@@ -143,7 +143,7 @@ private:
         stride_t stride_CbCr,
         stride_t stride_dst);
 
-    template <int incsp, int outcsp, int rgb_limit> void convert_translate_align(
+    template <uint64_t incsp, uint64_t outcsp, int rgb_limit> void convert_translate_align(
         const uint8_t* srcY,
         const uint8_t* srcCb,
         const uint8_t* srcCr,
@@ -154,7 +154,7 @@ private:
         stride_t stride_CbCr,
         stride_t stride_dst);
 
-    template <int incsp, int outcsp, int rgb_limit, int aligned> void convert_translate_dithering(
+    template <uint64_t incsp, uint64_t outcsp, int rgb_limit, int aligned> void convert_translate_dithering(
         const uint8_t* srcY,
         const uint8_t* srcCb,
         const uint8_t* srcCr,
@@ -165,7 +165,7 @@ private:
         stride_t stride_CbCr,
         stride_t stride_dst);
 
-    template <int incsp, int outcsp, int rgb_limit, int aligned, bool dithering> void convert_main(
+    template <uint64_t incsp, uint64_t outcsp, int rgb_limit, int aligned, bool dithering> void convert_main(
         const uint8_t* srcY,
         const uint8_t* srcCb,
         const uint8_t* srcCr,
@@ -176,7 +176,7 @@ private:
         stride_t stride_CbCr,
         stride_t stride_dst);
 
-    template <int incsp, int outcsp, int rgb_limit, int aligned, bool dithering> void convert_main_loop(
+    template <uint64_t incsp, uint64_t outcsp, int rgb_limit, int aligned, bool dithering> void convert_main_loop(
         const uint8_t* srcY,
         const uint8_t* srcCb,
         const uint8_t* srcCr,
@@ -189,7 +189,7 @@ private:
         int starty,
         int endy);
 
-    template<int incsp, int outcsp, int rgb_limit, int aligned, bool dithering> struct Tfunc_obj {
+    template<uint64_t incsp, uint64_t outcsp, int rgb_limit, int aligned, bool dithering> struct Tfunc_obj {
     private:
         const uint8_t* srcY;
         const uint8_t* srcCb;
