+++ src/ffmpeg/Tlibavcodec.h
@@ -49,7 +49,7 @@ public:
  static bool getVersion(const Tconfig *config,ffstring &vers,ffstring &license);
  static bool check(const Tconfig *config);
  static int lavcCpuFlags(void);
- static int ppCpuCaps(int csp);
+ static int ppCpuCaps(uint64_t csp);
  static void pp_mode_defaults(PPMode &ppMode);
  static int getPPmode(const TpostprocSettings *cfg,int currentq);
  static void swsInitParams(SwsParams *params,int resizeMethod);
